move send and receive streams into separate files
No functional change expected.
